requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langgraph
python-dotenv
sqlite-utils
